requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.6
django-filter==25.1
django-querycount==0.8.3
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
